- Moved trait of DAO access to core (APIWrapper) - Created a standalone class that implements with local DB - Changed return type of roomInfo from Future[Option[Room]] to Future[Room]   and return type of publicRoomInfo from Future[Unit] to Future[Room]
